add role selection in Thymleaf

+++ src/main/java/com/example/demo/repo/RoleRepo.java
@@ -2,5 +2,5 @@ package com.example.demo.repo;
 
 import com.example.demo.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface RoleRepo extends JpaRepository<Role, Integer> {
+public interface RoleRepo extends JpaRepository<Role, Long> {
 }
